backend/requirements.txt
attrs==20.3.0
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
ecdsa==0.14.1
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
httptools==0.1.1
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.3
python-jose==3.2.0
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.7
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
